textbox works, cannot submit
