+++ src/Components/Item.js
@@ -1,8 +1,20 @@
 import React from 'react';
 
-export const componentMapping = {
-  img: (content) => <Image {...content} />,
+// an array of objects to show each textarea or post on a page (
+// eslint-disable-next-line no-unused-vars
+const item = {
+  authorID: 'some-id',
+  postText: 'text',
+  postTitle: 'Title',
+  postLikes: 'postLikes',
+  userID: 'userID',
+  postComments: 'postComments',
+  type: 'some-type',
+  content: 'some-content',
 };
+// export const componentMapping = {
+//   img: (content) => <Image {...content} />,
+// };
 const Item = (props) => {
   const {
     type, content, updateItem, pressingKey,
@@ -25,6 +37,7 @@ const Item = (props) => {
           value={content}
           onChange={(e) => updateItem(e.target.value)}
         />
+      // eslint-disable-next-line no-undef
       ) : componentMapping[type](content)}
     </div>
   );
