+++ src/Components/Post.js
@@ -7,64 +7,52 @@ import React, {
 import Navigation from './Navigation';
 
 // this line will become const Post = (props) => { once there are props
-const Post = () => {
-  // eslint-disable-next-line no-unused-vars
-  const item = {
-    authorID: 'some-id',
-    postText: 'text',
-    postTitle: 'Title',
-    postLikes: 'postLikes',
-    userID: 'userID',
-    postComments: 'postComments',
-    type: 'type',
-  };
+const Post = () => (
+  <>
+    <Navigation />
+    <div>This is the discover page</div>
+    <div className="post">
+      <h3> Username </h3>
+      <div className="postTitle">
+        <h2>Post Title</h2>
+      </div>
+      <div className="postText">
+        <p>
+          Some information about some restaurant that user
+          has gone to and now talking about if friends should go
+          or not or if it really sucks.
 
-  return (
-    <>
-      <Navigation />
-      <div>This is the discover page</div>
-      <div className="post">
-        <h3> Username </h3>
-        <div className="postTitle">
-          <h2>Post Title</h2>
-        </div>
-        <div className="postText">
-          <p>
-            Some information about some restaurant that user
-            has gone to and now talking about if friends should go
-            or not or if it really sucks.
-
-          </p>
-          <div className="postComment">
-            <p>Some comment someone left about the post</p>
-          </div>
+        </p>
+        <div className="postComment">
+          <p>Some comment someone left about the post</p>
         </div>
       </div>
-    </>
-  );
+    </div>
+  </>
+);
 
-  // set state
-  // const [state, setState] = useState(value);
+// set state
+// const [state, setState] = useState(value);
 
-  // deconstruct props - uncomment the below
-  // once the props are used
-  // const [id, AuthorID, postText, postTitle, postLikes, userID, postComments] = props;
+// deconstruct props - uncomment the below
+// once the props are used
+// const [id, AuthorID, postText, postTitle, postLikes, userID, postComments] = props;
 
-  // TODO: fetch data from backend
-  //   useEffect(() => {
+// TODO: fetch data from backend
+//   useEffect(() => {
 
-  //   }, []);
+//   }, []);
 
-  //   // TODO: Render component
-  //   return (
-  //     <>
-  //       <div>this is a post</div>
-  //     </>
-  //   );
-  // };
+//   // TODO: Render component
+//   return (
+//     <>
+//       <div>this is a post</div>
+//     </>
+//   );
+// };
 
-  // TODO: PropTypes
-  // uncomment the props out once they are used
+// TODO: PropTypes
+// uncomment the props out once they are used
 //   Post.propTypes = {
 //     id: PropTypes.number.isRequired,
 //     AuthorID: PropTypes.string.isRequired,
@@ -80,5 +68,5 @@ const Post = () => {
 //     })).isRequired,
 //   };
 // };
-};
+
 export default Post;
